+++ src/Expression/ExpressionBuilderInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Expressions;
+namespace Yiisoft\Db\Expression;
 
 /**
  * Interface ExpressionBuilderInterface is designed to build raw SQL from specific expression objects that implement
