+++ src/Pdo/PdoValueBuilder.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pdo;
 
-use Yiisoft\Db\Expressions\ExpressionInterface;
-use Yiisoft\Db\Expressions\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\ExpressionInterface;
+use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 
 /**
  * Class PdoValueBuilder builds object of the {@see PdoValue} expression class.
