+++ src/Querys/Conditions/SimpleConditionBuilder.php
@@ -2,9 +2,9 @@
 
 namespace Yiisoft\Db\Querys\Conditions;
 
-use Yiisoft\Db\Expressions\ExpressionBuilderInterface;
-use Yiisoft\Db\Expressions\ExpressionBuilderTrait;
-use Yiisoft\Db\Expressions\ExpressionInterface;
+use Yiisoft\Db\Expression\ExpressionBuilderInterface;
+use Yiisoft\Db\Expression\ExpressionBuilderTrait;
+use Yiisoft\Db\Expression\ExpressionInterface;
 
 /**
  * Class NotConditionBuilder builds objects of {@see SimpleCondition}.
