+++ tests/ColumnSchemaBuilderTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Tests;
 
-use Yiisoft\Db\Expressions\Expression;
+use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Schemas\ColumnSchemaBuilder;
 use Yiisoft\Db\Schemas\Schema;
